Point 5: Remove Set from classname.
- Removed Set from SetCustomLoggingFilter and moved the class to its own module.
